fix: remove broken handlebar util package (#25)
Fun story, `handlebars-helpers` won't work because someone thought it
was a good idea to override the `require()` function, which breaks
`ncc`. And because it runs inline, it can't be conditionally not loaded.
:/

Anyway, I just wrote the one helper I needed inline. We can extend it as
needed.

Fixes common-config-elixir runs:
https://github.com/stordco/common-config-elixir/actions/runs/6343807016/job/17232517554